Reorg build scripts and travis.yml uploading to docker
