+++ BuildDocker.py
@@ -1,0 +1,3 @@
+from Utils import *
+
+buildDockerImage()
