+++ Deploy.sh
@@ -1,0 +1,3 @@
+docker login -u="$DOCKER_USERNAME" -p="$DOCKER_PASSWORD"
+docker build
+docker push
