+++ InstallConfigForDev.py
@@ -1,0 +1,6 @@
+from Utils import *
+	
+if(doesGitFolderExist()):
+	updateConfigs()
+else:
+	insertGitConfigSource()
