+++ build.py
@@ -1,18 +1,17 @@
-import subprocess
-import sys
+from Utils import *
 import os
+import sys
 
-sourcepath = "/src/UKSFWebsite.api/"
-buildpath = "./build_output"
-path = "/website-backend-config"
-dllpath = "./UKSFWebsite.api.dll"
-
-print("Build script starting in path - " + os.getcwd())
-
-
-try:
-   grepOut = subprocess.check_output(["dotnet", "publish", os.getcwd() + sourcepath, "-o", buildpath, "--framework", "netcoreapp1.0", "--runtime", sys.argv[1]])                      
-except subprocess.CalledProcessError as grepexc: 
-	print(grepexc.returncode)
-	print(grepexc.output)
-	sys.exit(grepexc.returncode)
+if(sys.argv(1) == "--noRebuild"):
+	os.chdir(buildpath)
+	startDotNetDll()
+elif(sys.argv(1) == "--onlyBuild"):
+	dotNetRestor()
+	tryPublish()
+else:
+	clearOldBuild()
+	dotNetRestor()
+	publishDotNetProgram()
+	insertGitConfigPublish()
+	os.chdir(buildpath)
+	startDotNetDll()
