--- getConfigs.py
@@ -1,7 +1,0 @@
-import os
-import sys
-import subprocess
-import shutil
-import sys
-
-subprocess.call(["git", "clone", "https://github.com/uksf/website-backend-config.git", sys.argv(1)])
