--- installDependencies.py
@@ -1,3 +1,0 @@
-import subprocess
-
-subprocess.call(["dotnet", "restore"])
